When the client disconnects while sending request body, most servers just start returning '' from env['wsgi.input'].read(..) and similar. This means the apps get truncated requests.
This commit makes req.body_file wrap the input stream in a proxy that detects
such conditions and raises a DisconnectError exception while reading. This is
used (and tested) to make sure req.body, req.POST, req.copy() and other request
methods and properties do not result in truncated request bodies or form
dictionaries, but raise an error instead.

The tests use actual wsgiref server to send truncated requests to and verify
that the codition is handled correctly in WebOb.