bring docs up-to-date with actual wsgify signature
the tests and code clearly expect (req, app, **kw) and it makes sense that it
would be that way since req is the ephemeral argument and the rest are
defined at setup-time. The documented order before was (app, req, **kw)
which puts the req in-between multiple config arguments in the
signature.

fixes #25, #120