Raise ValueError if cookie value has invalid chars
WebOb will now raise a ValueError if an attempt is made to set the
cookie to an invalid value. According to RFC6265 a cookie-octet has a
specific subset of allowed ASCII characters, and that subset does not
change whether the value is DQUOT'ed or not.

WebOb will still accept all other cookies, it just won't be able to
create them.