Add some tests for wsgify/wsgify.middleware args
When treating the result of `wsgify` or `wsgify.middleware` as a
callable, arguments passed to the args/kwargs options to wsgify (and, by
extension, to the middleware factory) are currently being dropped.

This adds a couple of tests to demonstrate this, as well as a couple of
tests to ensure that overriding args from the actual callable invocation
remains possible.

N.B. middleware only supports passing kwargs (and not args) through from
the intermediary stages (i.e. from unbound middleware -> middleware
factory -> app).