Correctly preserve args when using wsgify callables
wsgify objects can be created with args/kwargs at construction time:

    def hello(req, name):
        return "Hello, %s!" % name
    app = wsgify(hello, args=("Fred",))

Previously, these arguments would only be passed to the underlying
function (here, `hello`) when using the wsgify object as a callable that
accepts `(environ, start_response)`. If you used the object as a
callable taking a request (as documented) then the default args would be
discarded.

That is, continuing the above example:

    req = Request.blank('/')
    resp = req.get_response(app)  # => "Hello, Fred"

    BUT

    app(req)                      # => raises TypeError

This is perhaps most confusing when it interacts with
`wsgify.middleware`. For example:

     @wsgify.middleware
     def add_magic_header(req, app, value='seated ducks'):
         resp = req.get_response(app)
         resp.headers['Magic-Header'] = value
         return resp

     app = add_magic_header(app, value='flying elephants')

     resp1 = req.get_response(app)
     resp2 = app(req)

     resp1.headers['Magic-Header']  # => 'flying elephants'
     resp2.headers['Magic-Header']  # => 'seated ducks'

This commit ensures that if variadic arguments or keyword arguments are
not provided in a call like

    app(req)

then the defaults are used.