Add w.Accept.best_client_match
This flips the logic from best_match on it's head, in that it allows the
client to dictate what should be returned given a multitude of server
offerings.

With an accept header of:

application/json, text/html, */*;q=0.8

And a server offer of:

text/html, application/json

The best match returned will be 'application/json'. This makes a lot of
sense in other contexts too, for instance if the browser had a specific
ordering it wanted images in:

image/webp, image/jpeg, image/gif, image/*;q=0.9

If the server offered up image/gif, image/jpeg with best_match this
would return image/gif, with best_client_match, this would return
image/jpeg.