Change the default returned type to text/html
Currently if WebOb encounters a */* Accept header on an exception it
will return an application/json result. This leads to some interesting
results for crawlers and others that don't send an explicit HTTP Accept
and then end up receiving JSON instead of what was previously HTML.

This breaks B/W compat with issue #230.