Set content_type in headers only if we have one
It is entirely possible that if there is no content_type passed by the
user, no headerlist that contains a content type and there is no default
content type that we don't have a content_type.

In that case we want to make sure we don't set anything in the header
dictionary since None is not a valid header to return in a Response.