content_type removes charset unless text like
When changing the content_type, the charset parameter will no longer be
saved and transferred to the new content_type unless the new
content_type is likely to have charset as a parameter.

For now that includes text/*, application/xml, */*+xml. For any others
any change will require setting the charset explicitly after setting the
content_type.

Currently all other parameters are saved and restored onto the new
content_type, this is deprecated and will be removed in future versions,
and if parameters are to be kept, they will have to be kept explicitly.