Fix content_type bug if content_type was empty
When the new Content-Type value did not contain any semicolons we
attempted to save any existing parameters from the existing Content-Type
headers. If that header was empty, it would do nothing, and no
Content-Type would be set at all.

The new branch explicitly sets the Content-Type to the value if we don't
find any paramters to save.