Move the _lazified class out of the lazify function
Capturing the function, and returning a new object from a class that is
defined within the capturing function seems to cause some sort of really
pathological behavior that causes lazify() to take up a bunch of CPU
time, especially when it is part of a hot code path.

Thanks to Martijn Faassen for reporting!

Closes #282