Remove call to _check_offer().
There is no need to check offer. If we were to check, it would probably
be better as an assert, but there is no need, as the offers are
specified to be language tags in the documentation.

This method was copied from NilAccept.best_match. The original
corresponding method AcceptLanguage, AcceptLanguage.best_match, did not
call _check_offer either (not even in ._match).