Add __str__ and tests for the three AcceptLanguage classes.
Accept.__str__ returns a tidied up version of the header value, and
AcceptLanguageValidHeader.__str__ does the same.

NilAccept.__str__ returns ''; but given that the corresponding
Accept.__str__ returns a tidied up version of the header value,
returning '' would be misleading when the header is not empty, but
rather is not found in the request. We also need to return a different
value to indicate an invalid header, which cannot be parsed and tidied
up. So we instead remove the ambiguity with '<no header in request>' and
'<invalid header value>'.