Remove check for 'q'.
There is nothing in RFC 7231, section 5.3 (and 5.3.1 to 5.3.5) that says
any of

* media-range in Accept header
* charset in Accept-Charset header
* codings in Accept-Encoding header
* language-range in Accept-Language header

cannot be 'q'.

The only thing close to this is in section 5.3.2 "Accept":

'Although this prevents any media type parameter named "q" from being
used with a media range...'

but that says a media type parameter cannot be named 'q', not that a
media range cannot be 'q'. (`Accept.parse` currently ignores such
parameters, as mentioned in the module docstring.)

(The test for this was test_init_accept_invalid_value, removed in commit
b62746aa.)