Simplify handling of duplicate ranges in the header.
The RFC does not specify what to do when the same range appears in the
header more than once, possibly with different qvalues and giving
conflicting information. Previously, we chose which of the possibly
conflicting versions of the range to use for matching by first checking
if there was one with q=0, which would take priority; if one with q=0
was not found, we chose the one with the highest qvalue.

Having given it a lot more thought, it now seems clear that that kind of
special handling is not necessary -- the simplest way to handle it is to
take the first appearance of the range in the header (from the left) and
simply use that for matching.