MIMEAccept mimics old behaviour better
This updates MIMEAccept to mimic old behaviour better so that downstream
projects are not as broken.

The goal is not to provide the exact same behaviour, and thus MIMEAccept
still uses the new Accept header parsing and matching, however it is
much closer to the old API because it no longer raises on an invalid
header vs what was shipped in WebOb 1.8.0.