Restore _check_offer calls to best_match and __contains__.
This is so that best_match and __contains__ raise a ValueError with any
offer that contains '*'(s), as the methods did in WebOb 1.5, and so that
the behavior is consistent across AcceptValidHeader, AcceptNoHeader and
AcceptInvalidHeader.

The code supporting '*'s in offers has intentionally not been removed.

See https://github.com/Pylons/webob/issues/367.